Rarity Changes Item Stats
Added Update Stats method to items. converted all ints to doubles.

+++ TRP/Body.cs
@@ -7,21 +7,21 @@ namespace TRP
     class Body
     {
         protected string name;
-        protected int power;
+        protected double power;
 
         public string Name { get { return name; } set { name = value; } }
-        public int Power { get { return power; } set { power = value; } }
+        public double Power { get { return power; } set { power = value; } }
 
     }
 
     class Fighter : Body
     {
         protected List<Item> inventory = new List<Item>();
-        protected int hitPoints;
-        protected int attackPoints;
+        protected double hitPoints;
+        protected double attackPoints;
 
-        public int AttackPoints { get { return attackPoints; } set { attackPoints = value; } }
-        public int HitPoints { get { return hitPoints; } set { hitPoints = value; } }
+        public double AttackPoints { get { return attackPoints; } set { attackPoints = value; } }
+        public double HitPoints { get { return hitPoints; } set { hitPoints = value; } }
         public List<Item> Inventory
         {
             get { return inventory; }
@@ -33,7 +33,7 @@ namespace TRP
         private Weapon equippedWeapon = new Weapon("Sword", 4);
         public Weapon EquippedWeapon { get { return equippedWeapon; } set { equippedWeapon = value; } }
 
-        public Player(string name, int hitPoints, int power, Weapon equippedWeapon)
+        public Player(string name, double hitPoints, double power, Weapon equippedWeapon)
         {
             this.equippedWeapon = equippedWeapon;
             this.name = name;
@@ -74,7 +74,7 @@ namespace TRP
 
     class Monster : Fighter , ICloneable
     { 
-        public Monster(string name, int hitPoints, int attackPoints)
+        public Monster(string name, double hitPoints, double attackPoints)
         {
             this.name = name;
             this.hitPoints = hitPoints;
